refactor(security): move security configs to dedicated package
Rename McpServerConfiguration to HttpSecurityConfiguration and move both
HttpSecurityConfiguration and MethodSecurityConfiguration from
org.apache.solr.mcp.server.config to org.apache.solr.mcp.server.security.
