+++ src/main/java/org/apache/solr/mcp/server/security/HttpSecurityConfiguration.java
@@ -1,0 +1,83 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.apache.solr.mcp.server.security;
+
+import java.util.List;
+import org.springaicommunity.mcp.security.server.config.McpServerOAuth2Configurer;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Profile;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.annotation.web.configurers.CsrfConfigurer;
+import org.springframework.security.web.SecurityFilterChain;
+import org.springframework.web.cors.CorsConfiguration;
+import org.springframework.web.cors.CorsConfigurationSource;
+import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
+
+@Profile("http")
+@Configuration
+@EnableWebSecurity
+class HttpSecurityConfiguration {
+
+	@Value("${spring.security.oauth2.resourceserver.jwt.issuer-uri:}")
+	private String issuerUrl;
+
+	@Bean
+	@ConditionalOnProperty(name = "spring.security.enabled", havingValue = "true", matchIfMissing = true)
+	SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
+		return http
+				// ⬇️ Open every request on the server
+				.authorizeHttpRequests(auth -> {
+					auth.requestMatchers("/actuator").permitAll();
+					auth.requestMatchers("/actuator/*").permitAll();
+					auth.requestMatchers("/mcp").permitAll();
+					auth.anyRequest().authenticated();
+				})
+				// Configure OAuth2 on the MCP server
+				.with(McpServerOAuth2Configurer.mcpServerOAuth2(), (mcpAuthorization) -> {
+					// REQUIRED: the issuerURI
+					mcpAuthorization.authorizationServer(issuerUrl);
+				})
+				// MCP inspector
+				.cors(cors -> cors.configurationSource(corsConfigurationSource())).csrf(CsrfConfigurer::disable)
+				.build();
+	}
+
+	@Bean
+	@ConditionalOnProperty(name = "spring.security.enabled", havingValue = "false")
+	SecurityFilterChain unsecured(HttpSecurity http) throws Exception {
+		return http.authorizeHttpRequests(auth -> auth.anyRequest().permitAll())
+				// MCP inspector
+				.cors(cors -> cors.configurationSource(corsConfigurationSource())).csrf(CsrfConfigurer::disable)
+				.build();
+	}
+
+	public CorsConfigurationSource corsConfigurationSource() {
+		CorsConfiguration configuration = new CorsConfiguration();
+		configuration.setAllowedOriginPatterns(List.of("*"));
+		configuration.setAllowedMethods(List.of("*"));
+		configuration.setAllowedHeaders(List.of("*"));
+		configuration.setAllowCredentials(true);
+
+		UrlBasedCorsConfigurationSource source = new UrlBasedCorsConfigurationSource();
+		source.registerCorsConfiguration("/**", configuration);
+		return source;
+	}
+}
